word 模板

+++ solo-util-core/src/main/java/cn/sexycode/util/core/file/ZipUtils.java
@@ -1,6 +1,8 @@
 package cn.sexycode.util.core.file;
 
 import cn.sexycode.util.core.io.IOUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.io.*;
 import java.util.Enumeration;
@@ -14,6 +16,7 @@ import java.util.zip.ZipOutputStream;
  * @author qinzaizhen
  */
 public class ZipUtils {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ZipUtils.class);
 
     /**
      * @param replaceEntry    替换zip的文件路径以及对应的文件流
@@ -21,7 +24,7 @@ public class ZipUtils {
      * @param zipOutputStream 目标zip
      */
     public static void replaceItem(Map<String, InputStream> replaceEntry, ZipInputStream zipInputStream,
-            ZipOutputStream zipOutputStream) {
+            ZipOutputStream zipOutputStream) throws IOException {
         //
         ZipEntry entryIn;
         try {
@@ -35,6 +38,7 @@ public class ZipUtils {
                     if (in != null) {
                         // 使用替换流（替换文字内容）
                         zipOutputStream.write(IOUtils.toByteArray(in));
+                        in.close();
                     }
                 } else {
                     // 输出普通Zip流
@@ -45,7 +49,8 @@ public class ZipUtils {
                 zipOutputStream.closeEntry();
             }
         } catch (IOException e) {
-            e.printStackTrace();
+            LOGGER.error("失败", e);
+            throw e;
         } finally {
             close(zipInputStream);
             close(zipOutputStream);
@@ -59,33 +64,38 @@ public class ZipUtils {
      * @param destDirPath 解压后的目标文件夹
      * @throws RuntimeException 解压失败会抛出运行时异常
      */
-    public static void unZip(File srcFile, String destDirPath) throws RuntimeException {
+    public static void unZip(File srcFile, String destDirPath) {
         long start = System.currentTimeMillis();
         // 判断源文件是否存在
         if (!srcFile.exists()) {
             throw new RuntimeException(srcFile.getPath() + "所指文件不存在");
         }
         // 开始解压
-        ZipFile zipFile = null;
         try {
-            zipFile = new ZipFile(srcFile);
+            ZipFile zipFile = new ZipFile(srcFile);
             Enumeration<?> entries = zipFile.entries();
             while (entries.hasMoreElements()) {
                 ZipEntry entry = (ZipEntry) entries.nextElement();
-                System.out.println("解压" + entry.getName());
+                LOGGER.debug("解压" + entry.getName());
                 // 如果是文件夹，就创建个文件夹
                 if (entry.isDirectory()) {
                     String dirPath = destDirPath + "/" + entry.getName();
                     File dir = new File(dirPath);
-                    dir.mkdirs();
+                    if (!dir.exists() && !dir.mkdirs()) {
+                        throw new RuntimeException("解压失败, 无法创建目录" + dir.getAbsolutePath());
+                    }
                 } else {
                     // 如果是文件，就先创建一个文件，然后用io流把内容copy过去
                     File targetFile = new File(destDirPath + "/" + entry.getName());
                     // 保证这个文件的父文件夹必须要存在
                     if (!targetFile.getParentFile().exists()) {
-                        targetFile.getParentFile().mkdirs();
+                        if (!targetFile.getParentFile().mkdirs()) {
+                            throw new RuntimeException("解压失败, 无法创建目录" + targetFile.getParentFile().getAbsolutePath());
+                        }
                     }
-                    targetFile.createNewFile();
+                    if (!targetFile.exists() && !targetFile.createNewFile()) {
+                        throw new RuntimeException("解压失败, 无法创建文件" + targetFile.getAbsolutePath());
+                    }
                     // 将压缩文件内容写入到这个文件中
                     InputStream is = zipFile.getInputStream(entry);
                     FileOutputStream fos = new FileOutputStream(targetFile);
@@ -96,17 +106,10 @@ public class ZipUtils {
                 }
             }
             long end = System.currentTimeMillis();
-            System.out.println("解压完成，耗时：" + (end - start) + " ms");
+            LOGGER.debug("解压完成，耗时：" + (end - start) + " ms");
         } catch (Exception e) {
-            throw new RuntimeException("unzip error from ZipUtils", e);
-        } finally {
-            if (zipFile != null) {
-                try {
-                    zipFile.close();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
+            LOGGER.error("异常", e);
+            throw new RuntimeException("unzip error", e);
         }
     }
 
@@ -114,16 +117,14 @@ public class ZipUtils {
      * 包装输入流
      */
     public static ZipInputStream wrapZipInputStream(InputStream inputStream) {
-        ZipInputStream zipInputStream = new ZipInputStream(inputStream);
-        return zipInputStream;
+        return new ZipInputStream(inputStream);
     }
 
     /**
      * 包装输出流
      */
     public static ZipOutputStream wrapZipOutputStream(OutputStream outputStream) {
-        ZipOutputStream zipOutputStream = new ZipOutputStream(outputStream);
-        return zipOutputStream;
+        return new ZipOutputStream(outputStream);
     }
 
     private static void close(InputStream inputStream) {
